fix: Fixed neovim installer

+++ Makefile
@@ -36,8 +36,8 @@ PACKER=~/.local/share/nvim/site/pack/packer/start/packer.nvim
 
 install-neovim:
 	-[ -d $(NEOVIM) ] && mv $(NEOVIM) $(NEOVIM).backup
-	-[ ! -f $(PACKER) ] && git clone --depth 1 https://github.com/wbthomason/packer.nvim $(PACKER)
-	ln -s $(PWD)/nvim $(NEOVIM)
+	-[ ! -d $(PACKER) ] && git clone --depth 1 https://github.com/wbthomason/packer.nvim $(PACKER)
+	-[ ! -L $(NEOVIM) ] && ln -s $(PWD)/nvim $(NEOVIM)
 
 TMUX=~/.tmux.conf
 install-tmux:
